+++ R/size.R
@@ -15,7 +15,12 @@ NULL
 #' @param verbose A \code{logical} value, whether to output messages.
 #' @param ... Further arguments (used only for backwards compatibility).
 #' @rdname size-method
-#' @return An \code{integer} vector if s_attribute is \code{NULL}, a \code{data.table} otherwise.
+#' @return If \code{.Object} is a corpus (a \code{corpus} object or specified by
+#'   corpus id), an \code{integer} vector if argument \code{s_attribute} is
+#'   \code{NULL}, a two-column \code{data.table} otherwise (first column is the
+#'   s-attribute, second column: "size"). If \code{.Object} is a
+#'   \code{subcorpus_bundle} or a \code{partition_bundle}, a \code{data.table}
+#'   (with columns "name" and "size").
 #' @seealso See \code{\link{dispersion}}-method for counts of hits. The \code{\link{hits}}
 #' method calls the \code{size}-method to get sizes of subcorpora.
 #' @aliases size,slice-method
@@ -43,6 +48,10 @@ NULL
 #' size(sc, s_attribute = "speaker")
 #' size(sc, s_attribute = "party")
 #' size(sc, s_attribute = c("speaker", "party"))
+#' 
+#' # for subcorpus_bundle
+#' subcorpora <- corpus("GERMAPARLMINI") %>% split(s_attribute = "date")
+#' size(subcorpora)
 setGeneric("size", function(x, ...) UseMethod("size"))
 
 #' @rdname size-method
@@ -113,6 +122,20 @@ setMethod("size", "slice", function(x, s_attribute = NULL, ...){
 setMethod("size", "partition", function(x, s_attribute = NULL, ...){
   callNextMethod(x = x, s_attribute = s_attribute, ...)
 })
+
+#' @rdname size-method
+setMethod("size", "partition_bundle", function(x){
+  cpos_list <- lapply(x@objects, slot, "cpos")
+  cpos_matrix <- do.call(rbind, cpos_list)
+  dt <- data.table(cpos_matrix)
+  setnames(dt, old = c("V1", "V2"), new = c("cpos_left", "cpos_right"))
+  dt[, "name" := unlist(Map(f = rep, x = lapply(x@objects, slot, "name"), times = lapply(cpos_list, nrow)))]
+  dt[, "region_size" := dt[["cpos_right"]] - dt[["cpos_left"]] + 1L]
+  y <- dt[, sum(.SD[["region_size"]]), by = "name"]
+  setnames(y, old = "V1", new = "size")
+  y
+})
+
 
 #' @describeIn subcorpus Get the size of a \code{subcorpus} object from the
 #'   respective slot of the object.